Added a function capable of reading name and value of a button - not the best solution but works for now

+++ Thesis/AnimatedSprite.cpp
@@ -19,10 +19,10 @@ void AnimatedSprite::setFrame(size_t index, bool resetTime)
 		m_Vertices[2].position = sf::Vector2f(rect.width, rect.height);
 		m_Vertices[3].position = sf::Vector2f(rect.width, 0.f);
 
-		float left = rect.left;
-		float right = left + rect.width;
-		float top = rect.top;
-		float bottom = top + rect.height;
+		float left =   static_cast<float>(rect.left);
+		float right =  static_cast<float>(left + rect.width);
+		float top = static_cast<float>(rect.top);
+		float bottom = static_cast<float>(top + rect.height);
 
 		m_Vertices[0].texCoords = sf::Vector2f(left, top);
 		m_Vertices[1].texCoords = sf::Vector2f(left, bottom);
